dev commit

+++ contracts/nft/HeroNFTMSCombiner.sol
@@ -140,7 +140,7 @@ contract HeroNFTMSCombiner is
         uint32 spliteCount = uint32(costv >> 32 & 0xffffffff);
         require(spliteCount > 0, "HeroNFTMSCombiner: wrong shard splite count");
 
-        uint256 fuelCost = uint32(costv & 0xffffffff) * 10**18;
+        uint256 fuelCost = uint256(costv & 0xffffffff) * 10**18;
         if(fuelCost > 0){
             require(ERC20Burnable(_fuelTokenAddr).balanceOf(_msgSender()) >= fuelCost, "MysteryShard: insufficient fuel");
             ERC20Burnable(_fuelTokenAddr).burnFrom(_msgSender(), fuelCost);
@@ -165,7 +165,7 @@ contract HeroNFTMSCombiner is
         uint32 combineCount = uint32(costv >> 32 & 0xffffffff);
         require(combineCount > 0, "HeroNFTMSCombiner: wrong shard combine count");
 
-        uint256 fuelCost = uint32(costv & 0xffffffff) * 10**18;
+        uint256 fuelCost = uint256(costv & 0xffffffff) * 10**18;
         if(fuelCost > 0){
             require(ERC20Burnable(_fuelTokenAddr).balanceOf(_msgSender()) >= fuelCost, "MysteryShard: insufficient fuel");
             ERC20Burnable(_fuelTokenAddr).burnFrom(_msgSender(), fuelCost);
@@ -208,10 +208,10 @@ contract HeroNFTMSCombiner is
     }
     function _encodeShardId(ShardAttr memory attr) internal pure returns(uint256 tokenId) {
         tokenId = (0x01 << 64) | // token type = 1 : Mystery Shard
-            ((attr.shardID & 0xffff) << 48 ) |
-            ((attr.grade & 0xff) << 40 ) |
-            ((attr.shardType & 0xff) << 32 ) |
-            ((attr.randomType & 0xffff) << 16 ) |
-            (attr.mysteryType & 0xffff);
+            ((uint256(attr.shardID) & 0xffff) << 48 ) |
+            ((uint256(attr.grade) & 0xff) << 40 ) |
+            ((uint256(attr.shardType) & 0xff) << 32 ) |
+            ((uint256(attr.randomType) & 0xffff) << 16 ) |
+            (uint256(attr.mysteryType) & 0xffff);
     }
 }
